refactor: OpenAI configuration, controller, resdto

+++ src/main/java/com/voicecut/api/openai/dto/RiskAssessmentResponse.java
@@ -1,4 +1,4 @@
 package com.voicecut.api.openai.dto;
 
-public record RiskAssessmentResponse(double riskScore) {
+public record RiskAssessmentResponse(int riskScore) {
 }
